+++ src/app/components/ReactionList/index.tsx
@@ -10,23 +10,17 @@
 import React from 'react';
 import {TagFacesRounded} from '@material-ui/icons';
 import SettingsModal from "../SettingsModal";
-import {IMessage} from "../../repository/message/interface";
+import {IMessage, IReactionInfo} from "../../repository/message/interface";
 import i18n from "../../services/i18n";
-import APIManager from "../../services/sdk";
 import {InputPeer} from "../../services/sdk/messages/core.types_pb";
-import {ReactionList as RiverReactionList} from "../../services/sdk/messages/chat.messages_pb";
 import UserAvatar from "../UserAvatar";
 import UserName from "../UserName";
 import Chip from "@material-ui/core/Chip";
+import MessageRepo from "../../repository/message";
 import {Loading} from "../Loading";
-import {find} from "lodash";
 import {localize} from "../../services/utilities/localize";
 
 import './style.scss';
-
-interface IReactionInfo extends RiverReactionList.AsObject {
-    counter?: number;
-}
 
 interface IProps {
     onSelect?: (id: number, reaction: string, remove: boolean) => void;
@@ -40,7 +34,7 @@ interface IState {
 }
 
 class ReactionList extends React.Component<IProps, IState> {
-    private apiManager: APIManager;
+    private messageRepo: MessageRepo;
 
     constructor(props: IProps) {
         super(props);
@@ -52,29 +46,31 @@ class ReactionList extends React.Component<IProps, IState> {
             reactionList: [],
         };
 
-        this.apiManager = APIManager.getInstance();
+        this.messageRepo = MessageRepo.getInstance();
     }
 
     public openDialog(peer: InputPeer, message: IMessage) {
         this.setState({
+            loading: true,
             message,
             open: true,
-        });
-        this.apiManager.reactionList(peer, message.id || 0, 0).then((res) => {
-            (res.listList || []).map((item: IReactionInfo) => {
-                const t = find(message.reactionsList || [], {reaction: item.reaction});
-                if (t) {
-                    item.counter = t.total;
-                }
-                return item;
-            });
-            this.setState({
-                loading: false,
-                reactionList: res.listList || [],
-            });
-        }).catch(() => {
-            this.setState({
-                loading: false,
+            reactionList: (message.reactionsList || []).map((o) => {
+                return {
+                    counter: o.total,
+                    reaction: o.reaction,
+                    useridsList: [],
+                };
+            }),
+        }, () => {
+            this.messageRepo.getReactionList(peer, message.id || 0).then((res) => {
+                this.setState({
+                    loading: false,
+                    reactionList: res || [],
+                });
+            }).catch((err) => {
+                this.setState({
+                    loading: false,
+                });
             });
         });
     }
@@ -93,15 +89,14 @@ class ReactionList extends React.Component<IProps, IState> {
                                       label={localize(item.counter || 0)} className="reaction-chip"/>
                             </div>
                             <div className="reaction-user">
-                                {item.useridsList.map((id) => {
+                                {!loading ? item.useridsList.map((id) => {
                                     return (<Chip key={id} avatar={<UserAvatar id={id || '0'}/>}
                                                   tabIndex={-1} className="user-chip"
                                                   label={<UserName id={id || '0'} you={true} unsafe={true}/>}/>);
-                                })}
+                                }) : <Loading/>}
                             </div>
                         </div>);
                     })}
-                    {loading && <Loading/>}
                 </div>
             </SettingsModal>
         );
@@ -109,6 +104,7 @@ class ReactionList extends React.Component<IProps, IState> {
 
     private modalCloseHandler = () => {
         this.setState({
+            loading: false,
             message: {},
             open: false,
         });