+++ src/app/services/sdk/messages/chat.seal_pb.d.ts
@@ -1,16 +1,4 @@
 /* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
-/* tslint:disable */
 // package: msg
 // file: chat.seal.proto