+++ src/app/services/sdk/messages/chat.seal_pb.js
@@ -11,6 +11,7 @@
 /* eslint-disable */
 /* eslint-disable */
 /* eslint-disable */
+/* eslint-disable */
 // source: chat.seal.proto
 /**
  * @fileoverview